Updated to R-devel revision 72195.

+++ R/src/modules/internet/libcurl.c
@@ -153,23 +153,27 @@ static int curlMultiCheckerrs(CURLM *mhnd)
     for(int n = 1; n > 0;) {
 	CURLMsg *msg = curl_multi_info_read(mhnd, &n);
 	if (msg && (msg->data.result != CURLE_OK)) {
-	    const char *url, *strerr;
+	    const char *url, *strerr, *type;
 	    long status = 0;
 	    curl_easy_getinfo(msg->easy_handle, CURLINFO_EFFECTIVE_URL, &url);
 	    curl_easy_getinfo(msg->easy_handle, CURLINFO_RESPONSE_CODE,
 			      &status);
+	    // This reports the redirected URL
 	    if (status >= 400) {
-		if (url && url[0] == 'h')
+		if (url && url[0] == 'h') {
 		    strerr = http_errstr(status);
-		else
+		    type = "HTTP";
+		} else {
 		    strerr = ftp_errstr(status);
-		warning(_("URL '%s': status was '%d %s'"), url, status,
-			strerr);
+		    type = "FTP";
+		}
+		warning(_("cannot open URL '%s': %s status was '%d %s'"),
+			url, type, status, strerr);
 	    } else {
 		strerr = curl_easy_strerror(msg->data.result);
 		warning(_("URL '%s': status was '%s'"), url, strerr);
 	    }
-	    retval += 1;
+	    retval++;
 	}
     }
     return retval;
@@ -374,12 +378,15 @@ static
 int progress(void *clientp, double dltotal, double dlnow,
 	     double ultotal, double ulnow)
 {
+    CURL *hnd = (CURL *) clientp;
+    long status;
+    curl_easy_getinfo(hnd, CURLINFO_RESPONSE_CODE, &status);
+
     // we only use downloads.  dltotal may be zero.
-    if (dltotal > 0.) {
+    if ((status < 300) && (dltotal > 0.)) {
 	if (total == 0.) {
 	    total = dltotal;
 	    char *type = NULL;
-	    CURL *hnd = (CURL *) clientp;
 	    curl_easy_getinfo(hnd, CURLINFO_CONTENT_TYPE, &type);
 	    if (total > 1024.0*1024.0)
 		// might be longer than long, and is on 64-bit windows
@@ -394,15 +401,13 @@ int progress(void *clientp, double dltotal, double dlnow,
 	    if (R_Consolefile) fflush(R_Consolefile);
 	}
 	putdashes(&ndashes, (int)(50*dlnow/total));
-    } else {
-        total = 0.;             /* multiple hops on FOLLOWLOCATION */
     }
     return 0;
 }
-#endif
+#endif // _WIN32
 
 extern void Rsleep(double timeint);
-#endif
+#endif // HAVE_LIBCURL
 
 /* download(url, destfile, quiet, mode, headers, cacheOK) */
 
@@ -466,18 +471,18 @@ in_do_curlDownload(SEXP call, SEXP op, SEXP args, SEXP rho)
 	curl_easy_setopt(hnd[i], CURLOPT_TCP_KEEPALIVE, 1L);
 	if (!cacheOK)
 	    curl_easy_setopt(hnd[i], CURLOPT_HTTPHEADER, slist1);
-	curl_easy_setopt(hnd[i], CURLOPT_HEADER, 0L);
 
 	/* check that destfile can be written */
 	file = translateChar(STRING_ELT(sfile, i));
 	out[i] = R_fopen(R_ExpandFileName(file), mode);
-	curl_easy_setopt(hnd[i], CURLOPT_WRITEDATA, out[i]);
-	curl_multi_add_handle(mhnd, hnd[i]);
 	if (!out[i]) {
 	    n_err += 1;
 	    warning(_("URL %s: cannot open destfile '%s', reason '%s'"),
 		    url, file, strerror(errno));
-	    continue;
+	    if (nurls == 1) break; else continue;
+	} else {
+	    curl_easy_setopt(hnd[i], CURLOPT_WRITEDATA, out[i]);
+	    curl_multi_add_handle(mhnd, hnd[i]);
 	}
 
 	total = 0.;
@@ -522,6 +527,12 @@ in_do_curlDownload(SEXP call, SEXP op, SEXP args, SEXP rho)
 
 	if (!quiet) REprintf(_("Trying URL '%s'\n"), url);
     }
+    
+    if (n_err == nurls) {
+	// no dest files could be opened, so bail out
+	curl_multi_cleanup(mhnd);
+	return ScalarInteger(1);
+    }
 
     R_Busy(1);
     //  curl_multi_wait needs curl >= 7.28.0 .
@@ -558,9 +569,11 @@ in_do_curlDownload(SEXP call, SEXP op, SEXP args, SEXP rho)
     if (R_Consolefile) fflush(R_Consolefile);
 #endif
     if (nurls == 1) {
+	long status;
+	curl_easy_getinfo(hnd[0], CURLINFO_RESPONSE_CODE, &status);
 	double cl, dl;
 	curl_easy_getinfo(hnd[0], CURLINFO_SIZE_DOWNLOAD, &dl);
-	if (!quiet) {
+	if (!quiet && status == 200) {
 	    if (dl > 1024*1024)
 		REprintf(_("Downloaded %0.1f MB"), (double)dl/1024/1024);
 	    else if (dl > 10240)
@@ -578,23 +591,25 @@ in_do_curlDownload(SEXP call, SEXP op, SEXP args, SEXP rho)
 
     for (int i = 0; i < nurls; i++) {
 	if (out[i]) {
-            /* FIXME: remove empty / corrupt file on error: see below */
             fclose(out[i]);
+	    double dl;
+	    curl_easy_getinfo(hnd[i], CURLINFO_SIZE_DOWNLOAD, &dl);
+	    long status;
+	    curl_easy_getinfo(hnd[i], CURLINFO_RESPONSE_CODE, &status);
+	    // should we do something about incomplete transfers?
+	    if (status != 200 && dl == 0. && strchr(mode, 'w'))
+		unlink(R_ExpandFileName(translateChar(STRING_ELT(sfile, i))));
 	}
-
 	curl_multi_remove_handle(mhnd, hnd[i]);
 	curl_easy_cleanup(hnd[i]);
     }
     curl_multi_cleanup(mhnd);
     if (!cacheOK) curl_slist_free_all(slist1);
 
-    if (n_err != 0) {
-	if (strchr(mode, 'w')) {
-	    for (int i = 0; i < nurls; i++)
-		unlink(R_ExpandFileName(translateChar(STRING_ELT(sfile, i))));
-	}
-	error(_("cannot download all files"));
-    }
+    if(nurls > 1) {
+	if (n_err == nurls) error(_("cannot download any files"));
+	else if (n_err) warning(_("some files were not downloaded"));
+    } else if(n_err) error(_("download failed"));
 
     return ScalarInteger(0);
 #endif
@@ -765,7 +780,7 @@ static Rboolean Curl_open(Rconnection con)
 	n_err += fetchData(ctxt);
     if (n_err != 0) {
 	Curl_close(con);
-	error(_("cannot open connection"), n_err);
+	error(_("cannot open the connection"), n_err);
     }
 
     con->isopen = TRUE;
